1. Переместил все в папку handler 2. Удалил дублирующийся EventHandler 3. Удалил дублирующийся achievement.yaml, добавил отсутствующие title на achievement 4. В AchievementProgressService удалена readonly, потомучто в orelse создавалась новая транзакция 5. В application.yaml исправлен импорт конфига 6. Удален дублирующийся код в NiceGuyAchievementHandler 7. В RecommendationEventDto исправлены int на long

+++ src/main/java/faang/school/achievement/dto/RecommendationEventDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RecommendationEventDto {
-    private int authorId;
-    private int receiverId;
+    private long authorId;
+    private long receiverId;
     private String content;
 }
